Updated README.md
I am trying to add teehr readme documentation in CIROH DocuHub, but I came across a minor issue with the image path.  One of the images is not loading correctly when I add it in DocuHub since it uses relative path. So i adjusted the image path in the readme to use absolute URLs instead.